Split out `WorkingBeatmapCache` from `BeatmapManager`
